Use array Destructuring
By using array destructuring on useState we can pull out from the array the elements and store them in different variables. That will make code cleaner by assigning the current state snapshot and the function that updates the state to different variables and use that variables in the rest of the code.