Split state into multiple states
Why? Because you do not have to merge manually state when you just want to update part of the state. Wehn using useSate and want to update a part of the state you need to save the rest of the state as well. In order to avoid all this code we can have multiple states which can update only one part of the state each time when we need to without loosing the rest parts of the state.