Pass the state input data to other component in order to show them as a list
The state input data are passed to the Ingredients component through the onAddIngredient fucntion which is triggered on submit the form. It is being stored to state by replcing the latest state of ingredients with the new one. Then other components like the IngredientList have access to the ingredients state as is being passed to them as a prop.

The input data that we are getting from the