Use useEffect hook to fetch data form database
The useEffect is executed only AFTER every time the component rerenders. However by executing a function inside useEffect which updates the state and cause a rerender we will end up with an infinite loop.To avoid this we add an empty array as a second argument to the useEffect which will stop the infinite loop and the useEffect will only run once, after the first render cycle.So it will act like the ComponentDidMount.

The array is empty as we do not have any external dependencies in the function of the useEffect. All the variables that we use in there are created on the fly, they are not defined outside of the useEffect. If we had any dependencies we would add them in the array and the useEffect would be executed evrey time these dependencies are updated. So the useEffect would act like componentDidUpdate.