Fetch filtered data
By using the useEffect with the dependency array as a second argument which contains the  filtersState and the onLoadingIngredients function we control the useEffect execution to be done everytime these two dependencies change. So everytime the user types something and everytime the onLoadIngredients function that is passed to the parent component is changed.