Fix/bug infinite loop by using callBack() hook
In the search component, useEffect is using an array with dependencies. One dependency is the function onLoadIngredients() which is passedto the parent component search. If this function changes the useEffect hook in Search component is gonna be executed. Indeed the function is changing because everytime the parent componet Search rerenders the function is recreated and that means it is causing an infinite loop.

To prevent the infinite loop we use callBack() hook which allows you to wrap a function with it and it caches this function so that it survives rerender cycle, so when the component rerender the function will not be recreated so it will not be changed threfore the onLoadIngredients() that we are passing as dependency in the useEffect hook in SearchComponent will not change and will not cause the useEffect to be executed again.