Fix the issue were we are sending request for each keystroke
We do not want to send request to the server for each keystroke, because we are spamming the server.
We can set a timer and when the timer expires we check if the current input is the same as the one in the beginning of the timer and only if it is the same we know that there has been a pause and the user paused typing and we want to send the request.

To get the current value of the input we use the useRef() hook. This allows to create a reference and connect the input dom element with that reference by using ref property supported by React.

The variable inputRef should alse be included in the dependency array of useEffect to make sure that it will rerun if the inputRef changes.