Clean up timers to achieve memory efficient
We do have a bunch of timers for every keystroke, we want to  make sure we always clear the previous timer because is in no use anymore.We do not want old timers going on behind the scenes in memory instead we want to  have only one timer at a time.

To achieve that we use a cleanup function of useEffect and it runs after the second time the useEffect is executed. The first time the useEffect is executed this cleanup function is not executed. It is executed when useEffect  is executed for a second time, that means when one of the dependecies are changed. If there are no dependencies the cleanup function runs when the component gets unmounted.