Optimize the app by using useCallBack and React.memo
There is an extra render cyle of the ingredientsForm. The addIngredient function that is passed from t
he parent Ingredients component to IngredientForm component is recreated everytime the parent componen
t is rerender. To prevent this we wrap the function with the useCallBack which caches the function and
 is not getting rebuilt. Therefore the react.memo() in ingredientForm component detects that the funct
ion that is getting is the old one as it is not rebuilt and threfore it prevents the ingredientForm co
mponet from rerendering.