Use the custom hook to replace the request of Deleting an ingredient.
The perfect place to dispatch the action DELETE which deletes ingredients from the state after the sendRequest is executed, is the useEffect. UseEffect() runs after a component is rendered. So when the sendRequest of the hook is called the data of the state of the hook are changed that means that the component will be re rendered due to that change in state. That means that the sendRequest function is finished, the data of the state got updated, and so we can dispatch the DELETE action to update the ingredients state.